pyproject.toml
[project]
name = "alexandria"
version = "0.1.0"
description = ""
authors = [
    {name = "lcscosta",email = "lcscosta@disroot.org"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "uuid (>=1.30,<2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.packages]]
include = "backend"